services/main-app/pyproject.toml
[tool.poetry]
name = "kanban-main-app"
version = "0.1.0"
description = "Kanban Tracker Main Application"
authors = ["Boris <soskinboris@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.121.0"
uvicorn = "^0.38.0"
dynaconf = "^3.2.12"
sqlalchemy = "^2.0.44"
alembic = "^1.17.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.0"
passlib = "^1.7.4"
python-jose = "^3.5.0"
python-multipart = "^0.0.20"
bcrypt = "^5.0.0"
pydantic = {extras = ["email"], version = "^2.12.4"}
aiofiles = "^25.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.3"
mypy = "^1.18.2"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
httpx = "^0.28.1"
testcontainers = {extras = ["postgres"], version = "^4.13.3"}

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

# Конфигурация Ruff
[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Конфигурация MyPy
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

# Конфигурация Pytest
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
pythonpath = ["src"]